refactor: modularize tasks and add parser
