+++ au_b24/funcs/tasks/__init__.py
@@ -1,0 +1,4 @@
+from .get import get_tasks
+from .delete import delete_task
+from .add import add_task
+from .parse import parse_tasks
